RHI 16.3.a: backend-neutral rhi::Capabilities
First RHI extraction increment. Move the GPU capabilities struct out of Vulkan:
new src/rhi/Capabilities.hpp defines saida::rhi::Capabilities with no Vk types
(maxSamples is a plain uint32_t sample count). graphics/RenderCapabilities.hpp
becomes a using-alias shim so every consumer stays unchanged; VulkanDevice.hpp now
includes vulkan.h directly (it relied on the capabilities header for it). src/rhi/
Rhi.hpp anchors the compile-time backend selection for the types to come.

Desktop behaviour unchanged (the field had no external readers). Build + 14/14
tests green.

+++ src/graphics/VulkanDevice.cpp
@@ -323,7 +323,7 @@ void VulkanDevice::queryCapabilities() {
     if (props.apiVersion < api) api = props.apiVersion;
     capabilities_.apiVersion = api;
     capabilities_.discreteGpu = props.deviceType == VK_PHYSICAL_DEVICE_TYPE_DISCRETE_GPU;
-    capabilities_.maxSamples = maxUsableSampleCount();
+    capabilities_.maxSamples = static_cast<uint32_t>(maxUsableSampleCount());
 
     // Modern feature query (core Vulkan 1.1/1.2/1.3 feature structs). Only valid
     // to chain when the device exposes at least 1.3; otherwise leave caps false.
